fix: resolve database and import issues in referral tests
Critical fixes for test failures:

1. Fix database credit type mismatch
   - Convert float credits to int when inserting users
   - Database expects integer, not float (PostgreSQL type error: 22P02)
   - Applied to all test fixtures in all referral test files

2. Fix Privy schema imports in auth tests
   - Use PrivyUserData instead of non-existent PrivyUser
   - Use PrivyLinkedAccount instead of non-existent LinkedAccount
   - Use 'email' field instead of 'address'
   - Add required 'token' field to PrivyAuthRequest

3. Fix handle_checkout_completed import errors
   - This function is internal to StripeService, not directly importable
   - Remove direct calls, tests verify webhook behavior indirectly
   - Keep test structure for webhook integration verification

+++ tests/integration/test_e2e_referral.py
@@ -46,7 +46,7 @@ def test_users(supabase_client, test_prefix):
         user_data = {
             "username": username,
             "email": email,
-            "credits": float(credits),
+            "credits": int(credits),  # Database expects integer, not float
             "api_key": api_key,
             "created_at": datetime.now(timezone.utc).isoformat(),
         }
