+++ tests/integration/test_referral_database_integrity.py
@@ -38,7 +38,7 @@ def test_db_users(supabase_client, test_prefix):
         user_data = {
             "username": username,
             "email": email,
-            "credits": float(credits),
+            "credits": int(credits),  # Database expects integer, not float
             "api_key": api_key,
             "created_at": datetime.now(timezone.utc).isoformat(),
         }
